+++ tests/mocking.rs
@@ -1,7 +1,7 @@
 use core::time::Duration;
 
 use bevy::{prelude::*, time::TimeUpdateStrategy};
-use bevy_enhanced_input::prelude::*;
+use bevy_enhanced_input::{context::ExternallyMocked, prelude::*};
 use test_log::test;
 
 #[test]
@@ -130,6 +130,39 @@ fn manual() {
     assert!(!*action);
     assert_eq!(state, ActionState::None);
     assert_eq!(events, ActionEvents::COMPLETED);
+}
+
+#[test]
+fn external_mock() {
+    let mut app = App::new();
+    app.add_plugins((MinimalPlugins, EnhancedInputPlugin))
+        .add_input_context::<TestContext>()
+        .finish();
+
+    app.world_mut().spawn((
+        TestContext,
+        actions!(
+            TestContext[(
+                Action::<Test>::new(),
+                ExternallyMocked,
+                ActionMock::once(ActionState::Fired, true)
+            )]
+        ),
+    ));
+
+    app.update();
+
+    let mut actions = app
+        .world_mut()
+        .query::<(&Action<Test>, &ActionState, &ActionEvents)>();
+
+    let (&action, &state, &events) = actions.single(app.world()).unwrap();
+    assert!(
+        !*action,
+        "action shouldn't be updated because it marked as mocked externally"
+    );
+    assert_eq!(state, ActionState::None);
+    assert_eq!(events, ActionEvents::empty());
 }
 
 #[derive(Component)]