+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Added a component `ExternallyMocked` to exclude actions from being updated. These actions are updated manually by the user.
+
 ### Changed
 
 - Made `ActionTime::update` public